backend/requirements.txt
Flask
Flask_SQLAlchemy
Flask_CORS
yfinance
python-dotenv
requests